a list of all the variable names is still not stopping input values from being excluded. It mostly looks like the modal galleries, but there are others too, like yorder and showerror

+++ R/serverhelp_bookmarking.R
@@ -4,7 +4,7 @@ observe_excludebookmark <- function(react, inputstosave = NULL, session = getDef
     react,
     {   
     toExclude <- setdiff(names(session$input), c(""))
-    cat(toExclude, file = "./data/inputbookmarkexcludes.txt", sep ="\n", append = TRUE)
+    # cat(toExclude, file = "./data/inputbookmarkexcludes.txt", sep ="\n", append = TRUE)
     setBookmarkExclude(toExclude)
     },
     ignoreNULL = FALSE,
@@ -15,23 +15,27 @@ observe_excludebookmark <- function(react, inputstosave = NULL, session = getDef
 
 
 # refresh it when package is loaded
-if (isTRUE(getOption("shiny.testmode"))){
+# manually remove FALSE here to generate the list
+if (FALSE && isTRUE(getOption("shiny.testmode"))){
   inputidslist_filename <- "./data/inputidslist.txt"
   if (file.exists(inputidslist_filename)){file.remove(inputidslist_filename)}
 }
 
 
-appendinputids <- function(ids, nssep = "-"){
- if (isTRUE(getOption("shiny.testmode"))){
- showNotification("updating list of inputs")
- ids <- gsub(paste0(".*",nssep), "", ids)
- ids <- unique(ids)
- if (file.exists(inputidslist_filename)){
-   existingids <- readLines(inputidslist_filename)
-   ids <- unique(c(existingids, ids))
- }
- ids <- sort(ids)
- cat(ids, file = inputidslist_filename, append = FALSE, sep = "\n")
+appendinputids <- function(session = getDefaultReactiveDomain(), nssep = "-"){
+ if (FALSE && isTRUE(getOption("shiny.testmode"))){
+   observe({
+   showNotification("updating list of inputs")
+   ids <- names(session$input)
+   ids <- gsub(paste0(".*",nssep), "", ids)
+   ids <- unique(ids)
+   if (file.exists(inputidslist_filename)){
+     existingids <- readLines(inputidslist_filename)
+     ids <- unique(c(existingids, ids))
+   }
+   ids <- sort(ids)
+   cat(ids, file = inputidslist_filename, append = FALSE, sep = "\n")
+ })
  }
 }
